requirements.txt
appdirs==1.4.0
blinker==1.4
bson==0.4.2
Django==1.9.7
django-nose==1.4.5
djangorestframework==3.1.3
mongoengine==0.9.0
nose==1.3.7
packaging==16.8
psycopg2==2.6.2
pycrypto==2.6.1
pymongo==3.2.1
pymssql==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.1
requests==2.13.0
six==1.10.0
